README - finding unused sprites
In debug mode, when the engine is closed, the console will print all unused files in red. An unused file is a file that may not have been rendered so far in the last run. This means that you must run the scene and have the sprite render for this to happen. Some sprites can be false positives if they are hardcoded to render in game